feat(assemblyscript): create a deployable version
Given that `mint` is not part of the spec, it was not included in the
main contract. This makes sense for demonstration purposes, but for
demonstrating a live version of a Fungible Token and reporting how much
gas it uses, it does not work.

This adds minimal changes to the AssemblyScript contract to make it
useful once deployed.

* change `mint` to `init`, since it's only meant to be called once
* move to `main.ts`
* update tests to match